TransmitPortSettingを別ファイルへ分割
